Add files via upload
Project 2 Python:
For this project, they ask to create a program that manages to calculate the gallons used and the amount that must be paid with the meter readings that the user gives to the program. It is a program to help the user to identify how much water he uses and how much he will have to pay for it.
What my program does is receive the code to identify what is going to be calculated for the payment. Then it asks the user for the meter reading, the beginning meter, and the ending meter reading. With this, what the program does is subtract the ending meter reading from the beginning meter reading and then divide it by ten to get the water usage. Then the program uses the water usage to calculate the amount that must be paid depending on the user's code, whether it is residential, commercial, or industrial.